Rename api/user to userLookup
